fixed logout dialog bug and add claim airdrop end status and so on (#128)
* add airdrop

* add airdrop

* update promotion

* finish airdrop

* update promotion

* update project

* update

* update style

* update mobile candy

* update home style

* update tabbar choose dom

* update quests => promotions empty text

* fixed mobile dialog bug

* remove claim promotion failed error message

* remove claim promotion failed error message

* finish pc promotions

* finish pc promotions and update user quests

* add promotions empty status

* fixed mobile project style bug and pc vw style bug

* update mobile _tavern history path

* fixed logout dialog bug and add claim airdrop end status and so on

+++ src/assets/utils/tool/index.js
@@ -357,7 +357,7 @@ export const weiToEth = (value) => {
   } else {
     value = web3js.fromWei(value)
   }
-  return parseFloat(parseInt(value).toFixed(4))
+  return parseFloat(parseFloat(value).toFixed(4))
 }
 
 export const sliceStr = (str, { start = 0, end = 10 } = {}) => {
